Remove offset compensation for `relative` and `variadic` type identities
In Meevax, a reference to a bound variable is represented by de Bruijn
Index. If that reference is inserted by a syntactic closure, the
difference in scope between the macro definition environment and the
macro usage environment causes problems. If a bound variables identified
by considering only one of the environment, the resulting de Bruijn
Index will be incorrect.

The code removed by this commit tackles to fix this incorrect de Bruijn
Index. However, it did not consider the scope inside the macro
transformer, so it does not work correctly when the macro transformer
introduces new bindings inside it.

Until we have the desired solution to this problem, we will remove
support for inserting bound variables with syntactic closures. Along
with that, comment-out the test code that used the corresponding
feature.

Signed-off-by: yamacir-kit <httperror@404-notfound.jp>